preprod-eval/weather-chatbot/requirements.txt
openai
python-dotenv
azureml-core
azureml-dataprep
azureml-mlflow
azure-maps-search
pytest
pandas
xlsxwriter
mlflow
streamlit